feat: support speech_tokenizer_v3_25hz
- Added S3TokenizerV3, AudioEncoderV3, MultiHeadAttentionV3, ResidualAttentionBlockV3 in model_v3.py
- Implemented onnx2torch_v3 with bias detection for correct V3 conversion
- Updated LayerNorm to handle bfloat16 mixed precision on CPU
- Fixed apply_rotary_emb to avoid upcasting bfloat16
- Updated test_onnx.py to include v3 model tests with appropriate thresholds
- Updated README with V3 support info

+++ s3tokenizer/model_v2.py
@@ -55,8 +55,8 @@ def apply_rotary_emb(
 ) -> Tuple[torch.Tensor, torch.Tensor]:
     real = torch.view_as_real(freqs_cis)
     cos, sin = real[:, :, 0], real[:, :, 1]
-    cos = cos.unsqueeze(0).unsqueeze(2)
-    sin = sin.unsqueeze(0).unsqueeze(2)
+    cos = cos.unsqueeze(0).unsqueeze(2).to(xq.dtype)
+    sin = sin.unsqueeze(0).unsqueeze(2).to(xq.dtype)
 
     D = xq.shape[-1]
     half_l, half_r = xq[:, :, :, :D // 2], xq[:, :, :, D // 2:]
